change floating type in for loops of programs mpi_project_XR, angular_discrete_assign and angular_projection_matching

+++ src/xmipp/libraries/parallel/mpi_project_XR.cpp
@@ -76,7 +76,7 @@ void ProgMPIXrayProject::run()
     size_t idx = FIRST_IMAGE;
     size_t id;
 
-    for (double angle = projParam.tilt0; angle <= projParam.tiltF; angle += projParam.tiltStep)
+    for (int angle = projParam.tilt0; angle <= projParam.tiltF; angle += projParam.tiltStep)
     {
         if (projParam.singleProjection)
             data.fn_proj = projParam.fnOut;
